3/2 - 12 AM
Added a script that knows what the scene name is and changes the max speed of the car to expressway speeds for supercruise. Also added bounds on the sides of the blocks to block user from falling out of the map, and adjusted the time it takes to accelerate to max speed and brake to a complete stop for a more realistic feel may need to be adjusted for the build.