4/30
Hit boxes have been changed for ARB & CTA. They both work better now. The UI is scaled better but it is not the best. The way I made it is kinda hard to make it perfect. But it's better than before. Also I made the main menu have a black background. You can change this by changing settings in the Main Camera.